PORTALS-2566: refactor T&C link into its own component
