+++ apps/SageAccountWeb/src/components/ProfileValidation/TermsAndConditionsWrapped.tsx
@@ -1,11 +1,10 @@
-import { Box, Button } from '@mui/material'
+import { Box } from '@mui/material'
 import TermsAndConditions from 'synapse-react-client/dist/containers/TermsAndConditions'
-
 import React from 'react'
 import { ContinueButton } from './ContinueButton'
 import { VerificationSubmission } from 'synapse-react-client/dist/utils/synapseTypes'
 import { ReturnToAppButton } from './ReturnToAppButton'
-import { URL_TERMS_CONDITIONS_AGREEMENT } from 'synapse-react-client/dist/utils/SynapseConstants'
+import { TermsAndConditionsLink } from 'components/TermsAndConditionsLink'
 
 export type TermsAndConditionsWrappedProps = {
   verificationSubmission: VerificationSubmission
@@ -30,14 +29,9 @@ const TermsAndConditionsWrapped: React.FC<TermsAndConditionsWrappedProps> = ({
         disabled={!isFormComplete}
         onClick={() => onNext(verificationSubmission)}
       />
-      <Button
-        variant="text"
-        href={URL_TERMS_CONDITIONS_AGREEMENT}
-        target="_blank"
+      <TermsAndConditionsLink
         sx={{ width: '100%', padding: '6px', marginTop: '10px' }}
-      >
-        View and Complete Terms and Conditions for Use
-      </Button>
+      />
       <ReturnToAppButton />
     </Box>
   )
